Changed Queue_lib.h file: removed implementation of function. Added *.c file for functions implementation.
